Updated base project structure

+++ source/Window.d
@@ -1,17 +1,12 @@
 module d2d.Window;
 
-import d2d.event.Keyboard;
-import d2d.event.Mouse;
+import d2d.Keyboard;
+import d2d.Mouse;
 
 class Window {
 
-    private double framerate;
-    //private SDL_Window* window;
-    //private SDL_Renderer* renderer;
-    //private Mix_Music music;
+    private int framerate;
     private Keyboard keyboard;
     private Mouse mouse;
-
-    //TODO: Main loop updates mouse and keyboard and present renderer.
 
 }
